refactor(ui): extract loader-reveal timing into a shared utility
Move the delayed-reveal / minimum-visible-time logic out of App.vue's
displayApp() into utils/loaderReveal.ts so it can be imported as-is by
ui-ee via the existing "kestra/src/*" alias, instead of duplicating it.
