+++ ui/src/App.vue
@@ -39,18 +39,7 @@
     import DrillDownDrawer from "./components/dashboard/DrillDownDrawer.vue"
     import PwaInstallPrompt from "./components/PwaInstallPrompt.vue"
     import {useThemeCycle} from "./composables/useThemeCycle"
-
-    declare global {
-        interface Window {
-            __kestraLoader?: {
-                shown: boolean
-                shownAt: number
-                showDelay: number
-                minVisible: number
-                timer: ReturnType<typeof setTimeout> | null
-            }
-        }
-    }
+    import {revealApp} from "./utils/loaderReveal"
 
     const loaded = ref(false)
 
@@ -100,30 +89,7 @@
         Utils.switchTheme(miscStore)
         applyFontScale(getAppFontSizeMode())
 
-        const loader = document.getElementById("loader-wrapper")
-        const appContainer = document.getElementById("app-container")
-        const kestraLoader = window.__kestraLoader
-
-        const reveal = () => {
-            if (loader) {
-                loader.classList.remove("is-visible")
-                // let the fade-out transition finish before removing it from layout
-                setTimeout(() => { loader.style.display = "none" }, 150)
-            }
-            if (appContainer) appContainer.style.display = "block"
-            loaded.value = true
-        }
-
-        if (kestraLoader?.timer) {
-            clearTimeout(kestraLoader.timer)
-        }
-
-        if (kestraLoader?.shown) {
-            const elapsed = performance.now() - kestraLoader.shownAt
-            setTimeout(reveal, Math.max(0, kestraLoader.minVisible - elapsed))
-        } else {
-            reveal()
-        }
+        revealApp(() => { loaded.value = true })
     }
 
     watch(() => route?.meta?.anonymous, async (anonymous) => {
